+++ net/wire/msgmempool_test.go
@@ -22,7 +22,7 @@ func TestMemPool(t *testing.T) {
 	}
 
 	// Ensure max payload is expected value.
-	wantPayload := uint32(0)
+	wantPayload := uint64(0)
 	maxPayload := msg.MaxPayloadLength(pver)
 	if maxPayload != wantPayload {
 		t.Errorf("MaxPayloadLength: wrong max payload length for "+
