+++ util/varbytes.go
@@ -6,13 +6,13 @@ import (
 	"io"
 )
 
-func ReadVarBytes(r io.Reader, maxAllowed uint32, fieldName string) ([]byte, error) {
+func ReadVarBytes(r io.Reader, maxAllowed uint64, fieldName string) ([]byte, error) {
 	count, err := ReadVarInt(r)
 	if err != nil {
 		return nil, err
 	}
 
-	if count > uint64(maxAllowed) {
+	if count > maxAllowed {
 		str := fmt.Sprintf("%s is larger that the max allowed size count %d,max %d", fieldName, count, maxAllowed)
 		return nil, errors.New(str)
 	}
